break up distirbutions in single files
