+++ lib/python/picongpu/picmi/__init__.py
@@ -8,9 +8,10 @@ from .solver import ElectromagneticSolver
 from .gaussian_laser import GaussianLaser
 from .species import Species
 from .layout import PseudoRandomLayout
-from .distribution import UniformDistribution
-from .distribution import FoilDistribution
 from . import constants
+
+from .distribution import FoilDistribution
+from .distribution import UniformDistribution
 
 import picmistandard
 
@@ -25,8 +26,8 @@ __all__ = [
     "GaussianLaser",
     "Species",
     "PseudoRandomLayout",
-    "UniformDistribution",
     "FoilDistribution",
+    "UniformDistribution",
     "constants",
 ]
