+++ lib/python/picongpu/picmi/distribution/__init__.py
@@ -1,0 +1,11 @@
+"""
+PICMI for PIConGPU
+"""
+
+from .UniformDistribution import UniformDistribution
+from .FoilDistribution import FoilDistribution
+
+__all__ = [
+    "UniformDistribution",
+    "FoilDistribution",
+]
